move notificationSound.mp3 to assets folder to be accessible to webpack
